+++ src/components/sporsmal/Sporsmal.tsx
@@ -20,21 +20,4 @@ const RadioPG = () => {
   );
 };
 
-/* function RadioPG() : ReactElement {
-  return (
-    <div className="radioContainer" key="melk">
-      <input
-        type="radio"
-        id="melk"
-        name="melk"
-        value="melk"
-        checked= {true}
-        className="skjemaelement__label"
-      />
-
-      <label className= "skjemaelement__label" htmlFor= "Ja-svar">Ja</label>
-    </div>
-  );
-} */
-
 export default RadioPG;
