+++ src/components/sporsmal/sporsmal.less
