+++ src/constants/sideIDKonstanter.ts
@@ -1,11 +1,12 @@
 export const sideHjelpeteksterID = {
   UTBETALING: 'UTBETALING',
   DAGENS_TRANSPORTMIDDEL: 'DAGENS_TRANSPORTMIDDEL',
-  REISETILSKUDD_PERIODE_TRANSPORTMIDDEL: 'REISETILSKUDD_PERIODE_TRANSPORTMIDDEL',
+  REISETILSKUDD_PERIODE: 'REISETILSKUDD_PERIODE',
 };
 
 export const sideHjelpetekster = {
   UTBETALING: 'Utbetaling av pengene',
   DAGENS_TRANSPORTMIDDEL: 'Dagens transportmiddel',
-  REISETILSKUDD_PERIODE_TRANSPORTMIDDEL: 'Transportmiddel i reisetilskuddperioden',
+  REISETILSKUDD_PERIODE: 'Transportmiddel i reisetilskuddperioden',
 };
+
