+++ src/pages/dagens-transportmiddel/dagens-transportmiddel.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const DagensTransportmiddel = () => {
+    return(
+        <div>
+
+        </div>
+    )
+}
+
+export default DagensTransportmiddel;
