+++ src/pages/kvittering-side/kvittering-side.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const KvitteringSide = () => {
+    return(
+        <div>
+
+        </div>
+    )
+}
+
+export default KvitteringSide;
