+++ src/pages/reisetilskudd-periode/reisetilskudd-periode.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Filopplaster from '../../components/filopplaster/Filopplaster';
+
+const ReiseTilskuddPeriode = () => {
+    return(
+        <div>
+            <Filopplaster
+                tillatteFiltyper={['image/png', 'image/jpeg']}
+                maxFilstørrelse={1024 * 1024}
+            />
+        </div>
+    )
+}
+
+export default ReiseTilskuddPeriode;
