+++ src/pages/utbetaling/utbetaling.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { Undertittel } from 'nav-frontend-typografi';
+import RadioPG from '../../components/sporsmal/Sporsmal';
+
+const Utbetaling = () => {
+    return(
+        <div>
+            <Undertittel>
+                Utbetaling
+            </Undertittel>
+            <RadioPG />
+        </div>
+    )
+}
+
+export default Utbetaling;
